refactor: Move functions from FileWriter.test.ts to FileWriter.ts
