+++ tests/lib/FileWriter.test.ts
@@ -1,18 +1,4 @@
-const newLineChar = '\n';
-function appendToEndOfFile(initialContent: string, textToAppend: string) {
-    if (textToAppend.length === 0) {
-        return initialContent;
-    }
-    let result = initialContent;
-    if (result.length > 0 && !result.endsWith(newLineChar)) {
-        result += newLineChar;
-    }
-    result += textToAppend;
-    if (!textToAppend.endsWith(newLineChar)) {
-        result += newLineChar;
-    }
-    return result;
-}
+import { appendToEndOfFile, appendToListWithinFile, newLineChar } from '../../src/lib/FileWriter';
 
 describe('FileWriter', () => {
     it('should be able to append to an empty file', () => {
@@ -133,26 +119,3 @@ Sed ipsam libero qui consequuntur quaerat non atque quia ab praesentium explicab
         expect(newFile2).toEqual(expectedContent);
     });
 });
-
-function appendToListWithinFile(initialContent: string, targetListHeading: string, textToAppend: string) {
-    if (textToAppend.length === 0) {
-        return initialContent;
-    }
-    if (targetListHeading === '') {
-        throw Error('Cannot move line to list as empty target list heading was supplied');
-    }
-    const linesArray = initialContent.split('\n');
-    const headingLine = linesArray.indexOf(targetListHeading);
-
-    if (headingLine === -1) {
-        let result = initialContent;
-        if (result.length > 0 && !result.endsWith(newLineChar)) {
-            result += newLineChar;
-        }
-        result += targetListHeading + newLineChar + textToAppend + newLineChar;
-        return result;
-    } else {
-        linesArray[headingLine] += newLineChar + textToAppend;
-        return linesArray.join(newLineChar);
-    }
-}
